Step 5: Migrate scrapers to modular structure with backward compat
- Created src/dtu_analyzer/scrapers/async_scraper.py
  - Moved from root scraper_async.py
  - Updated imports to use config, logger, and parser modules
  - Removed duplicate utility functions
  - Uses config.paths for file locations

- Created src/dtu_analyzer/scrapers/threaded_scraper.py
  - Moved from root scraper.py
  - Updated imports to use config, logger, and parser modules
  - extractGrades and extractReviews now use shared parsers
  - Uses config.paths for file locations

- Created backward-compatible wrappers:
  - scraper_async.py → imports from new location
  - scraper.py → imports from new location

- Updated src/dtu_analyzer/scrapers/__init__.py
  - Exports async_main and threaded_main

All existing scripts and workflows continue to work via wrappers
